Create Fluid (Texatherm 32)
Hello everyone,
For my final paper I have to work with CoolProp in Python. I should use texatherm 32 as a fluid. I don't know how to add or create it.
On line 39 it says I use water, but this should be Texatherm 32. Texatherm 22 is in there but so I need texatherm 32....
Can anyone help me with this?
Thanks in advance